7장-연습 . DateIndex 파싱 통과 1

+++ src/test/java/cahhp07/practice/ColumnsIndexInfo.java
@@ -2,14 +2,14 @@ package cahhp07.practice;
 
 public class ColumnsIndexInfo {
 //    String defaultColumns = "Dimension.DATE,Ad unit 1,Ad unit 2,Ad unit ID 1,Ad unit ID 2,Column.TOTAL_AD_REQUESTS,Column.TOTAL_LINE_ITEM_LEVEL_IMPRESSIONS,Column.TOTAL_LINE_ITEM_LEVEL_CLICKS,Column.TOTAL_LINE_ITEM_LEVEL_CPM_AND_CPC_REVENUE,Column.TOTAL_ACTIVE_VIEW_VIEWABLE_IMPRESSIONS_RATE";
-    private byte dateIndex;
-    private byte unitNameIndex;
-    private byte unitIdIndex;
-    private byte requestIndex;
-    private byte impressionIndex;
-    private byte clickIndex;
-    private byte revenueIndex;
-    private byte viewablityIndex;
+    private int dateIndex;
+    private int unitNameIndex;
+    private int unitIdIndex;
+    private int requestIndex;
+    private int impressionIndex;
+    private int clickIndex;
+    private int revenueIndex;
+    private int viewablityIndex;
 
     @Override
     public String toString() {
@@ -25,67 +25,67 @@ public class ColumnsIndexInfo {
                 '}';
     }
 
-    public byte getDateIndex() {
+    public int getDateIndex() {
         return dateIndex;
     }
 
-    public void setDateIndex(byte dateIndex) {
+    public void setDateIndex(int dateIndex) {
         this.dateIndex = dateIndex;
     }
 
-    public byte getUnitNameIndex() {
+    public int getUnitNameIndex() {
         return unitNameIndex;
     }
 
-    public void setUnitNameIndex(byte unitNameIndex) {
+    public void setUnitNameIndex(int unitNameIndex) {
         this.unitNameIndex = unitNameIndex;
     }
 
-    public byte getUnitIdIndex() {
+    public int getUnitIdIndex() {
         return unitIdIndex;
     }
 
-    public void setUnitIdIndex(byte unitIdIndex) {
+    public void setUnitIdIndex(int unitIdIndex) {
         this.unitIdIndex = unitIdIndex;
     }
 
-    public byte getRequestIndex() {
+    public int getRequestIndex() {
         return requestIndex;
     }
 
-    public void setRequestIndex(byte requestIndex) {
+    public void setRequestIndex(int requestIndex) {
         this.requestIndex = requestIndex;
     }
 
-    public byte getImpressionIndex() {
+    public int getImpressionIndex() {
         return impressionIndex;
     }
 
-    public void setImpressionIndex(byte impressionIndex) {
+    public void setImpressionIndex(int impressionIndex) {
         this.impressionIndex = impressionIndex;
     }
 
-    public byte getClickIndex() {
+    public int getClickIndex() {
         return clickIndex;
     }
 
-    public void setClickIndex(byte clickIndex) {
+    public void setClickIndex(int clickIndex) {
         this.clickIndex = clickIndex;
     }
 
-    public byte getRevenueIndex() {
+    public int getRevenueIndex() {
         return revenueIndex;
     }
 
-    public void setRevenueIndex(byte revenueIndex) {
+    public void setRevenueIndex(int revenueIndex) {
         this.revenueIndex = revenueIndex;
     }
 
-    public byte getViewablityIndex() {
+    public int getViewablityIndex() {
         return viewablityIndex;
     }
 
-    public void setViewablityIndex(byte viewablityIndex) {
+    public void setViewablityIndex(int viewablityIndex) {
         this.viewablityIndex = viewablityIndex;
     }
 }
